requirements.txt
# Dependencias Python para el Proyecto
paho-mqtt>=1.6.1
requests>=2.28.0
